TKSS-1229: Backport JDK-8368301: sun/security/util/math/intpoly compiler warnings

+++ kona-crypto/src/main/java/com/tencent/kona/sun/security/util/math/intpoly/Curve448OrderField.java
@@ -36,7 +36,7 @@ public final class Curve448OrderField extends IntegerPolynomial {
     private static final int MAX_ADDS = 1;
     public static final BigInteger MODULUS = evaluateModulus();
     private static final long CARRY_ADD = 1 << 27;
-    private static final int LIMB_MASK = -1 >>> (64 - BITS_PER_LIMB);
+    private static final long LIMB_MASK = -1L >>> (64 - BITS_PER_LIMB);
 
     public static final Curve448OrderField ONE = new Curve448OrderField();
